==> Analysis in R/imcRtools/4_Spatial_Analysis_clinical2.R <==


#3- Analysis using imcRtools functions
#The Bodernmiller's pipeline works with both, SPE and SCE, just need to figure out the correct slots of SCE that should be used in the functions.
#Have a look at the help of each function to check which slot of the SCE should be used.

setwd("/Users/joaoluizsfilho/Library/CloudStorage/Dropbox/Work_Files/Matthias_Lab/Projects/COSMX_colon_dataset/COSMX_Colonrectal_cancer_project/TMA_original/imcrtools")


#3.1- Testing some plotting from Bodernmiller's tutorial using the SingleCellExperiment
library(imcRtools)
library(cytomapper)
library(openxlsx)
library(stringr)
library(dittoSeq)
library(RColorBrewer)
library(Rphenograph)
library(igraph)
library(dittoSeq)
library(viridis)
library(bluster)
library(BiocParallel)
library(ggplot2)
library(scran)
library(CATALYST)
library(kohonen)
library(ConsensusClusterPlus)
library(patchwork)
library(pheatmap)
library(gridExtra)
library(SingleCellExperiment)
library(SpatialExperiment)
library(tidyverse)
library(ggridges)
library(scater)
library(cowplot)
library(viridis)
library(dplyr)
library(readr)

# 1- Read in the spatial experiment object
# 1.1- steinbock generated data
ad_ep1 <- readRDS("./1_RDS_files/ad_ep1_sce.rds")
ad_ep2 <- readRDS("./1_RDS_files/ad_ep2_sce.rds")
ad_ep3 <- readRDS("./1_RDS_files/ad_ep3_sce.rds")

ad_tme1 <- readRDS("./1_RDS_files/ad_tme1_sce.rds")
ad_tme2 <- readRDS("./1_RDS_files/ad_tme2_sce.rds")
ad_tme3 <- readRDS("./1_RDS_files/ad_tme3_sce.rds")

ad_sma1 <- readRDS("./1_RDS_files/ad_sma1_sce.rds")
ad_sma2 <- readRDS("./1_RDS_files/ad_sma2_sce.rds")
ad_sma3 <- readRDS("./1_RDS_files/ad_sma3_sce.rds")

ad_cms1 <- readRDS("./1_RDS_files/ad_cms1_sce.rds")
ad_cms2 <- readRDS("./1_RDS_files/ad_cms2_sce.rds")
ad_cms3 <- readRDS("./1_RDS_files/ad_cms3_sce.rds")
ad_cms4 <- readRDS("./1_RDS_files/ad_cms4_sce.rds")
ad_cms5 <- readRDS("./1_RDS_files/ad_cms5_sce.rds")

ad_dmmr <- readRDS("./1_RDS_files/ad_dmmr_sce.rds")
ad_pmmr <- readRDS("./1_RDS_files/ad_pmmr_sce.rds")
ad_partial <- readRDS("./1_RDS_files/ad_partial_sce.rds")


# Adding spatial local and global data in colData for each sce object
ad_norecurrence <- readRDS("./1_RDS_files/ad_norecurrence_sce.rds")
ad_local <- readRDS("./1_RDS_files/ad_local_sce.rds")
ad_liver <- readRDS("./1_RDS_files/ad_liver_sce.rds")
ad_brain <- readRDS("./1_RDS_files/ad_brain_sce.rds")
ad_multisite <- readRDS("./1_RDS_files/ad_multisite_sce.rds")

spatial_local <- ad_multisite@int_colData@listData[["reducedDims"]]@listData[["spatial"]]
spatial_local <- as.data.frame(spatial_local)

spatial_global <- ad_multisite@int_colData@listData[["reducedDims"]]@listData[["spatial_fov"]]
spatial_global <- as.data.frame(spatial_global)

CenterX_local_px <- spatial_local$V1
CenterY_local_px <- spatial_local$V2
CenterX_global_px <- spatial_global$V1
CenterY_global_px <- spatial_global$V2

ad_multisite$CenterX_local_px <- CenterX_local_px
ad_multisite$CenterY_local_px <- CenterY_local_px
ad_multisite$CenterX_global_px <- CenterX_global_px
ad_multisite$CenterY_global_px <- CenterY_global_px

saveRDS(ad_norecurrence, "./1_RDS_files/ad_norecurrence_sce.rds")
saveRDS(ad_local, "./1_RDS_files/ad_local_sce.rds")
saveRDS(ad_liver, "./1_RDS_files/ad_liver_sce.rds")
saveRDS(ad_brain, "./1_RDS_files/ad_brain_sce.rds")
saveRDS(ad_multisite, "./1_RDS_files/ad_multisite_sce.rds")

#By default, single-cell data is read in as SpatialExperiment object. 
#The summarized pixel intensities per channel and cell (here mean intensity) are stored in the counts slot. 
#Columns represent cells and rows represent channels.
counts(spe)[1:5,1:5]

#Metadata associated to individual cells are stored in the colData slot. 
#After initial image processing, these metadata include the numeric identifier (ObjectNumber), 
#the area, and morphological features of each cell. 
#In addition, sample_id stores the image name from which each cell was extracted and the width and height of the corresponding images are stored.
head(colData(sce))

#The main difference between the SpatialExperiment and the SingleCellExperiment data container in the current setting is the way spatial locations of all cells are stored. 
#For the SingleCellExperiment container, the locations are stored in the colData slot 
#while the SpatialExperiment container stores them in the spatialCoords slot:
head(spatialCoords(sce))

#The spatial object graphs generated by steinbock are read into a colPair slot of the SpatialExperiment (or SingleCellExperiment) object. 
#Cell-cell interactions (cells in close spatial proximity) are represented as “edge list” (stored as SelfHits object). 
#Here, the left side represents the column indices of the “from” cells and the right side represents the column indices of the “to” cells.
colPair(sce, "neighborhood")

#Finally, metadata regarding the channels are stored in the rowData slot. 
#This information is extracted from the panel.csv file. 
#Channels are ordered by isotope mass and therefore match the channel order of the multi-channel images
head(rowData(spe))


#2- Single-cell processing
#After reading in the single-cell data, few further processing steps need to be taken.

#2.1- Add aditional metadata
# We can set the colnames of the object to generate unique identifiers per cell:
colnames(spe) <- paste0(spe$ROI, "_", spe$Master_Index)
colnames(spe)

#It is also often the case that sample-specific metadata are available externally. 
#For the current data, we need to link the cancer type (also referred to as “Indication”) to each sample. 
#This metadata is available as external excel file:
#library(openxlsx)
#library(stringr)
#meta <- read.xlsx("steinbock_outputs/sample_metadata.xlsx")
#spe$patient_id <- as.vector(str_extract_all(spe$sample_id, "Patient[1-9]", simplify = TRUE))
#spe$ROI <- as.vector(str_extract_all(spe$sample_id, "00[1-3]", simplify = TRUE))
#spe$indication <- meta$Indication[match(spe$patient_id, meta$Sample.ID)]

unique(spe$Group)
unique(spe$ROI)
unique(spe$CaseID)

#2.2- Transform counts - similar to quantile normalization
#The distribution of expression counts across cells is often observed to be skewed towards the right side meaning lots of cells display low counts and few cells have high counts.
#To avoid analysis biases from these high-expressing cells, the expression counts are commonly transformed or clipped.
#Here, we perform counts transformation using an inverse hyperbolic sine function. This transformation is commonly applied to flow cytometry data. 
#The cofactor here defines the expression range on which no scaling is performed. 
#While the cofactor for CyTOF data is often set to 5, IMC data usually display much lower counts. We therefore apply a cofactor of 1.
#However, other transformations such as log(counts(spe) + 0.01) should be tested when analysing IMC data.

library(dittoSeq)
dittoRidgePlot(spe, var = "CD3", group.by = "CaseID", assay = "counts") +
  ggtitle("CD3 - before transformation")

#the transformation step
assay(spe, "exprs") <- asinh(counts(spe)/1)

dittoRidgePlot(spe, var = "CD3", group.by = "patient_id", assay = "exprs") +
  ggtitle("CD3 - after transformation")

# 2.3- Define interesting channels
#For downstream analysis such as visualization, dimensionality reduction and clustering, only a subset of markers should be used. 
#As convenience, we can store an additional entry in the rowData slot that specifies the markers of interest. 
#Here, we deselect the nuclear markers, which were primarily used for cell segmentation, and keep all other biological targets.

rowData(spe)$use_channel <- !grepl("DNA|Histone", rownames(spe))

# 2.4- Define color schemes
#We will define color schemes for different metadata entries of the data and conveniently store them 
#in the metadata slot of the SpatialExperiment which will be helpful for downstream data visualizations. 
#We will use colors from the RColorBrewer and dittoSeq package but any other coloring package will suffice.

library(RColorBrewer)
color_vectors <- list()

ROI <- setNames(dittoColors(reps = 1)[seq_along(unique(spe$sample_id))], 
                unique(spe$sample_id))
Group <- setNames(brewer.pal(length(unique(spe$Group)), name = "Set2"), 
                  unique(spe$Group))

color_vectors$ROI <- ROI
color_vectors$Group <- Group
color_vectors
metadata(spe)$color_vectors <- color_vectors

unique(spe$cell_type)
unique(spe$hierarchy)

celltype <- setNames(c("#3F1B03", "#F4AD31", "#894F36", "#1C750C", "#EF8ECC", 
                       "#6471E2", "#4DB23B", "#F4800C", "#BF0A3D", "#000"),
                     c("B cells", "CD4 cells", "CD8 cells", "Epithelial", "Endothelium", 
                       "Macrophage", "Neutrophil", "Smooth Muscle Cell", "RBC", "B cell"))

metadata(spe)$color_vectors$celltype <- celltype

majorcelltype <- setNames(c("#3F1B03", "#6471E2", "#4DB23B", "#BF0A3D"),
                          c("Stromal", "Lymphoid", "Myeloid", "Vascular"))
unique(spe$pheno_cluster)

cellcluster <- setNames(c("#00924C","#00EA7B","#D5FFEB","#783D0D",
                              "#BC9201",'#93B220',"#6FB19B", "#D4E8E1",
                              "#AD0791","#FCCDE5","#BBABFF","#7853FF",
                              "#F62ED5","#FEFA35","#E42520","#FB9797",
                              "#FDC77F","#013F89","#01A6C7","#FF7F00",
                              "#5DA5FD","#7AE8FE"
),
c("Tumor epithelial cell", "ChemokinesHigh Tumor epithelial cell", "IGF2+AREG+ Tumor epithelial cell", "Stem cell", 
  "IGF2+AREG+ stem cell", "Goblet cell", "Tuft cell", "ChemokinesHigh tuft cell",
  "Plasma cell", 'B cell', "CD8+ T cell", "CD4+ T cell", "Treg",
  'Fibroblast', 'Pericyte', "Endothelial cell", 'Myofibroblast',
  'SPP1+ macrophage', 'SPP1+ monocyte', 'Neutrophil', 'ChemokinesHigh macrophage',
  'ChemokinesHigh monocyte'))

metadata(sce)$color_vectors$cellcluster <- cellcluster

##################################################################################################################################################################################################################################################################################################################################################################################################################################################################

# 1- Build the Neighborhood graph

# he buildSpatialGraph function operates on individual images. Therefore the returned object is grouped by entries in img_id. 
# This means all cells of a given image are grouped together in the object. 
# The ordering of cells within each individual image is the same as the ordering of these cells in the input object.

sce <- buildSpatialGraph(sce, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_global_px", "CenterY_global_px"))
colPair(spe, "expansion_interaction_graph")

ad_ep1 <- buildSpatialGraph(ad_ep1, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_global_px", "CenterY_global_px"))
colPair(ad_ep1, "expansion_interaction_graph")
ad_ep2 <- buildSpatialGraph(ad_ep2, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_global_px", "CenterY_global_px"))
ad_ep3 <- buildSpatialGraph(ad_ep3, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_global_px", "CenterY_global_px"))

ad_tme1 <- buildSpatialGraph(ad_tme1, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_global_px", "CenterY_global_px"))
ad_tme2 <- buildSpatialGraph(ad_tme2, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_global_px", "CenterY_global_px"))
ad_tme3 <- buildSpatialGraph(ad_tme3, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_global_px", "CenterY_global_px"))

ad_sma1 <- buildSpatialGraph(ad_sma1, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_global_px", "CenterY_global_px"))
ad_sma2 <- buildSpatialGraph(ad_sma2, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_global_px", "CenterY_global_px"))
ad_sma3 <- buildSpatialGraph(ad_sma3, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_global_px", "CenterY_global_px"))

ad_cms1 <- buildSpatialGraph(ad_cms1, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_global_px", "CenterY_global_px"))
ad_cms2 <- buildSpatialGraph(ad_cms2, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_global_px", "CenterY_global_px"))
ad_cms3 <- buildSpatialGraph(ad_cms3, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_global_px", "CenterY_global_px"))
ad_cms4 <- buildSpatialGraph(ad_cms4, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_global_px", "CenterY_global_px"))
ad_cms5 <- buildSpatialGraph(ad_cms5, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_global_px", "CenterY_global_px"))

ad_dmmr <- buildSpatialGraph(ad_dmmr, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_global_px", "CenterY_global_px"))
ad_pmmr <- buildSpatialGraph(ad_pmmr, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_global_px", "CenterY_global_px"))
ad_partial <- buildSpatialGraph(ad_partial, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_global_px", "CenterY_global_px"))

# test with local px
ad_norecurrence <- buildSpatialGraph(ad_norecurrence, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_local_px", "CenterY_local_px"))
ad_local <- buildSpatialGraph(ad_local, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_local_px", "CenterY_local_px"))
ad_liver <- buildSpatialGraph(ad_liver, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_local_px", "CenterY_local_px"))
ad_brain <- buildSpatialGraph(ad_brain, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_local_px", "CenterY_local_px"))
ad_multisite <- buildSpatialGraph(ad_multisite, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_local_px", "CenterY_local_px"))

# test with global px
ad_norecurrence <- buildSpatialGraph(ad_norecurrence, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_global_px", "CenterY_global_px"))
ad_local <- buildSpatialGraph(ad_local, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_global_px", "CenterY_global_px"))
ad_liver <- buildSpatialGraph(ad_liver, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_global_px", "CenterY_global_px"))
ad_brain <- buildSpatialGraph(ad_brain, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_global_px", "CenterY_global_px"))
ad_multisite <- buildSpatialGraph(ad_multisite, img_id = "sample_ID", type = "expansion", threshold = 60, coords = c("CenterX_global_px", "CenterY_global_px"))


# expansion interaction graph 
plotSpatial(sce, 
            node_color_by = "CIPR_ordered_clusters", 
            img_id = "sample_ID", 
            coords = c("CenterX_global_px", "CenterY_global_px"),
            draw_edges = TRUE, 
            colPairName = "expansion_interaction_graph", 
            node_size_fix = 1,
            nodes_first = FALSE, 
            directed = TRUE,
            edge_color_fix = "black") + 
  scale_color_manual(values = metadata(sce)$color_vectors$cellcluster)

plotSpatial(sce[,sce$sample_ID == "fov10"], 
            node_color_by = "CIPR_ordered_clusters", 
            img_id = "fov", 
            coords = c("CenterX_global_px", "CenterY_global_px"),
            draw_edges = TRUE, 
            colPairName = "expansion_interaction_graph", 
            node_size_fix = 1,
            nodes_first = FALSE, 
            directed = TRUE,
            edge_color_fix = "black") + 
  scale_color_manual(values = metadata(sce)$color_vectors$cellcluster) +
  ggtitle("fov23, Radius graph")

plotSpatial(sce[,sce$sample_ID == "fov14"], 
            node_color_by = "CIPR_ordered_clusters", 
            img_id = "fov", 
            coords = c("CenterX_global_px", "CenterY_global_px"),
            draw_edges = TRUE, 
            colPairName = "expansion_interaction_graph", 
            node_size_fix = 1,
            nodes_first = FALSE, 
            directed = TRUE,
            edge_color_fix = "black") + 
  scale_color_manual(values = metadata(sce)$color_vectors$cellcluster) +
  ggtitle("fov14, Radius graph")

spe_Pneumo <- buildSpatialGraph(spe_Pneumo, img_id = "ROI", type = "expansion", threshold = 20, coords = c("x", "y"))
colPair(spe_Pneumo, "expansion_interaction_graph")

spe_NonPneumo <- buildSpatialGraph(spe_NonPneumo, img_id = "ROI", type = "expansion", threshold = 20, coords = c("x", "y"))
colPair(spe_NonPneumo, "expansion_interaction_graph")

spe_HIVPOS <- buildSpatialGraph(spe_HIVPOS, img_id = "ROI", type = "expansion", threshold = 20, coords = c("x", "y"))
colPair(spe_HIVPOS, "expansion_interaction_graph")

spe_HIVNEG <- buildSpatialGraph(spe_HIVNEG, img_id = "ROI", type = "expansion", threshold = 20, coords = c("x", "y"))
colPair(spe_HIVNEG, "expansion_interaction_graph")

plotSpatial(spe_Pneumo, 
            node_color_by = "pheno_cluster", 
            img_id = "ROI", 
            coords = c("x", "y"),
            draw_edges = TRUE, 
            colPairName = "expansion_interaction_graph", 
            node_size_fix = 0.1,
            nodes_first = FALSE, 
            directed = FALSE,
            edge_color_fix = "grey") + 
  scale_color_manual(values = metadata(spe_Pneumo)$color_vectors$cellcluster) +
  ggtitle("Pneumonia - Radius graph")

plotSpatial(spe_NonPneumo, 
            node_color_by = "pheno_cluster", 
            img_id = "ROI", 
            coords = c("x", "y"),
            draw_edges = TRUE, 
            colPairName = "expansion_interaction_graph", 
            node_size_fix = 0.1,
            nodes_first = FALSE, 
            directed = FALSE,
            edge_color_fix = "grey") + 
  scale_color_manual(values = metadata(spe_NonPneumo)$color_vectors$cellcluster) +
  ggtitle("Non-Pneumonia - Radius graph")

plotSpatial(spe_HIVPOS, 
            node_color_by = "pheno_cluster", 
            img_id = "ROI", 
            coords = c("x", "y"),
            draw_edges = TRUE, 
            colPairName = "expansion_interaction_graph", 
            node_size_fix = 0.1,
            nodes_first = FALSE, 
            directed = FALSE,
            edge_color_fix = "grey") + 
  scale_color_manual(values = metadata(spe_HIVPOS)$color_vectors$cellcluster) +
  ggtitle("HIV POS - Radius graph")

##################################################################################################################################################################################################################################################################################################################################################################################################################################################################

# 2- Patch analysis T cells

# Epithelial clusters
ad_ep1 <- patchDetection(ad_ep1, patch_cells = ad_ep1$metaclusters == "T cell", img_id = "sample_ID", expand_by = 60, min_patch_size = 3, #change this
                       coords = c("CenterX_global_px", "CenterY_global_px"), colPairName = "expansion_interaction_graph")
ad_ep2 <- patchDetection(ad_ep2, patch_cells = ad_ep2$metaclusters == "T cell", img_id = "sample_ID", expand_by = 60, min_patch_size = 3, #change this
                       coords = c("CenterX_global_px", "CenterY_global_px"), colPairName = "expansion_interaction_graph")
ad_ep3 <- patchDetection(ad_ep3, patch_cells = ad_ep3$metaclusters == "T cell", img_id = "sample_ID", expand_by = 60, min_patch_size = 3, #change this
                       coords = c("CenterX_global_px", "CenterY_global_px"), colPairName = "expansion_interaction_graph")


patch_size <- patchSize(ad_ep1, "patch_id", coords = c("CenterX_global_px", "CenterY_global_px"))
patch_size <- merge(patch_size, colData(ad_ep1)[match(patch_size$patch_id, ad_ep1$patch_id),], by = "patch_id")

patch_size2 <- patchSize(ad_ep2, "patch_id", coords = c("CenterX_global_px", "CenterY_global_px"))
patch_size2 <- merge(patch_size2, colData(ad_ep2)[match(patch_size2$patch_id, ad_ep2$patch_id),], by = "patch_id")

patch_size3 <- patchSize(ad_ep3, "patch_id", coords = c("CenterX_global_px", "CenterY_global_px"))
patch_size3 <- merge(patch_size3, colData(ad_ep3)[match(patch_size3$patch_id, ad_ep3$patch_id),], by = "patch_id")

plotSpatial(ad_ep2, node_color_by = "patch_id", img_id = "sample_ID", node_size_fix = 1, coords = c("CenterX_global_px", "CenterY_global_px")) + theme(legend.position = "none") + scale_color_manual(values = col_PD)
plotSpatial(ad_ep3, node_color_by = "patch_id", img_id = "sample_ID", node_size_fix = 1, coords = c("CenterX_global_px", "CenterY_global_px")) + theme(legend.position = "none") + scale_color_manual(values = col_PD)

patch_size <- as.data.frame(patch_size)
patch_size2 <- as.data.frame(patch_size2)
patch_size3 <- as.data.frame(patch_size3)

patch_size_merge <- rbind(patch_size, patch_size2, patch_size3)

ggplot(patch_size_merge) + 
  geom_boxplot(aes(Epithelial_cluster, log10(size))) +
  geom_point(aes(Epithelial_cluster, log10(size))) + theme(axis.text.x = element_text(angle=90))

library(dplyr)
library(readr)
patch_size_merge <- patch_size_merge %>% select_if(~ !any(is.na(.)))
write_csv(patch_size_merge,"Epithelial_clusters_Tcell_patchsize.csv")


# TME clusters
ad_tme1 <- patchDetection(ad_tme1, patch_cells = ad_tme1$metaclusters == "T cell", img_id = "sample_ID", expand_by = 60, min_patch_size = 3, #change this
                       coords = c("CenterX_global_px", "CenterY_global_px"), colPairName = "expansion_interaction_graph")
ad_tme2 <- patchDetection(ad_tme2, patch_cells = ad_tme2$metaclusters == "T cell", img_id = "sample_ID", expand_by = 60, min_patch_size = 3, #change this
                       coords = c("CenterX_global_px", "CenterY_global_px"), colPairName = "expansion_interaction_graph")
ad_tme3 <- patchDetection(ad_tme3, patch_cells = ad_tme3$metaclusters == "T cell", img_id = "sample_ID", expand_by = 60, min_patch_size = 3, #change this
                       coords = c("CenterX_global_px", "CenterY_global_px"), colPairName = "expansion_interaction_graph")

patch_size <- patchSize(ad_tme1, "patch_id", coords = c("CenterX_global_px", "CenterY_global_px"))
patch_size <- merge(patch_size, colData(ad_tme1)[match(patch_size$patch_id, ad_tme1$patch_id),], by = "patch_id")

patch_size2 <- patchSize(ad_tme2, "patch_id", coords = c("CenterX_global_px", "CenterY_global_px"))
patch_size2 <- merge(patch_size2, colData(ad_tme2)[match(patch_size2$patch_id, ad_tme2$patch_id),], by = "patch_id")

patch_size3 <- patchSize(ad_tme3, "patch_id", coords = c("CenterX_global_px", "CenterY_global_px"))
patch_size3 <- merge(patch_size3, colData(ad_tme3)[match(patch_size3$patch_id, ad_tme3$patch_id),], by = "patch_id")

plotSpatial(ad_tme1, node_color_by = "patch_id", img_id = "sample_ID", node_size_fix = 1, coords = c("CenterX_global_px", "CenterY_global_px")) + theme(legend.position = "none") + scale_color_manual(values = col_PD)
plotSpatial(ad_tme2, node_color_by = "patch_id", img_id = "sample_ID", node_size_fix = 1, coords = c("CenterX_global_px", "CenterY_global_px")) + theme(legend.position = "none")
plotSpatial(ad_tme3, node_color_by = "patch_id", img_id = "sample_ID", node_size_fix = 1, coords = c("CenterX_global_px", "CenterY_global_px")) + theme(legend.position = "none") + scale_color_manual(values = col_PD)

patch_size <- as.data.frame(patch_size)
patch_size2 <- as.data.frame(patch_size2)
patch_size3 <- as.data.frame(patch_size3)

patch_size_merge <- rbind(patch_size, patch_size2, patch_size3)

ggplot(patch_size_merge) + 
  geom_boxplot(aes(TME_cluster, log10(size))) +
  geom_point(aes(TME_cluster, log10(size))) + theme(axis.text.x = element_text(angle=90))

patch_size_merge <- patch_size_merge %>% select_if(~ !any(is.na(.)))
write_csv(patch_size_merge,"TME_clusters_Tcell_patchsize.csv")


# SMA clusters
ad_sma1 <- patchDetection(ad_sma1, patch_cells = ad_sma1$metaclusters == "T cell", img_id = "sample_ID", expand_by = 60, min_patch_size = 3, #change this
                          coords = c("CenterX_global_px", "CenterY_global_px"), colPairName = "expansion_interaction_graph")
ad_sma2 <- patchDetection(ad_sma2, patch_cells = ad_sma2$metaclusters == "T cell", img_id = "sample_ID", expand_by = 60, min_patch_size = 3, #change this
                          coords = c("CenterX_global_px", "CenterY_global_px"), colPairName = "expansion_interaction_graph")
ad_sma3 <- patchDetection(ad_sma3, patch_cells = ad_sma3$metaclusters == "T cell", img_id = "sample_ID", expand_by = 60, min_patch_size = 3, #change this
                          coords = c("CenterX_global_px", "CenterY_global_px"), colPairName = "expansion_interaction_graph")

patch_size <- patchSize(ad_sma1, "patch_id", coords = c("CenterX_global_px", "CenterY_global_px"))
patch_size <- merge(patch_size, colData(ad_sma1)[match(patch_size$patch_id, ad_sma1$patch_id),], by = "patch_id")

patch_size2 <- patchSize(ad_sma2, "patch_id", coords = c("CenterX_global_px", "CenterY_global_px"))
patch_size2 <- merge(patch_size2, colData(ad_sma2)[match(patch_size2$patch_id, ad_sma2$patch_id),], by = "patch_id")

patch_size3 <- patchSize(ad_sma3, "patch_id", coords = c("CenterX_global_px", "CenterY_global_px"))
patch_size3 <- merge(patch_size3, colData(ad_sma3)[match(patch_size3$patch_id, ad_sma3$patch_id),], by = "patch_id")

plotSpatial(ad_sma1, node_color_by = "patch_id", img_id = "sample_ID", node_size_fix = 1, coords = c("CenterX_global_px", "CenterY_global_px")) + theme(legend.position = "none") + scale_color_manual(values = col_PD)
plotSpatial(ad_sma2, node_color_by = "patch_id", img_id = "sample_ID", node_size_fix = 1, coords = c("CenterX_global_px", "CenterY_global_px")) + theme(legend.position = "none") + scale_color_manual(values = col_PD)
plotSpatial(ad_sma3, node_color_by = "patch_id", img_id = "sample_ID", node_size_fix = 1, coords = c("CenterX_global_px", "CenterY_global_px")) + theme(legend.position = "none") + scale_color_manual(values = col_PD)

patch_size <- as.data.frame(patch_size)
patch_size2 <- as.data.frame(patch_size2)
patch_size3 <- as.data.frame(patch_size3)

patch_size_merge <- rbind(patch_size, patch_size2, patch_size3)

ggplot(patch_size_merge) + 
  geom_boxplot(aes(aSMA_cluster, log10(size))) +
  geom_point(aes(aSMA_cluster, log10(size))) + theme(axis.text.x = element_text(angle=90))

patch_size_merge <- patch_size_merge %>% select_if(~ !any(is.na(.)))
write_csv(patch_size_merge,"SMA_clusters_Tcell_patchsize.csv")

# CMS clusters
ad_cms1 <- patchDetection(ad_cms1, patch_cells = ad_cms1$metaclusters == "T cell", img_id = "sample_ID", expand_by = 60, min_patch_size = 3, #change this
                          coords = c("CenterX_global_px", "CenterY_global_px"), colPairName = "expansion_interaction_graph")
ad_cms2 <- patchDetection(ad_cms2, patch_cells = ad_cms2$metaclusters == "T cell", img_id = "sample_ID", expand_by = 60, min_patch_size = 3, #change this
                          coords = c("CenterX_global_px", "CenterY_global_px"), colPairName = "expansion_interaction_graph")
ad_cms3 <- patchDetection(ad_cms3, patch_cells = ad_cms3$metaclusters == "T cell", img_id = "sample_ID", expand_by = 60, min_patch_size = 3, #change this
                          coords = c("CenterX_global_px", "CenterY_global_px"), colPairName = "expansion_interaction_graph")
ad_cms4 <- patchDetection(ad_cms4, patch_cells = ad_cms4$metaclusters == "T cell", img_id = "sample_ID", expand_by = 60, min_patch_size = 3, #change this
                          coords = c("CenterX_global_px", "CenterY_global_px"), colPairName = "expansion_interaction_graph")
ad_cms5 <- patchDetection(ad_cms5, patch_cells = ad_cms5$metaclusters == "T cell", img_id = "sample_ID", expand_by = 60, min_patch_size = 3, #change this
                          coords = c("CenterX_global_px", "CenterY_global_px"), colPairName = "expansion_interaction_graph")

patch_size <- patchSize(ad_cms1, "patch_id", coords = c("CenterX_global_px", "CenterY_global_px"))
patch_size <- merge(patch_size, colData(ad_cms1)[match(patch_size$patch_id, ad_cms1$patch_id),], by = "patch_id")

patch_size2 <- patchSize(ad_cms2, "patch_id", coords = c("CenterX_global_px", "CenterY_global_px"))
patch_size2 <- merge(patch_size2, colData(ad_cms2)[match(patch_size2$patch_id, ad_cms2$patch_id),], by = "patch_id")

patch_size3 <- patchSize(ad_cms3, "patch_id", coords = c("CenterX_global_px", "CenterY_global_px"))
patch_size3 <- merge(patch_size3, colData(ad_cms3)[match(patch_size3$patch_id, ad_cms3$patch_id),], by = "patch_id")

patch_size4 <- patchSize(ad_cms4, "patch_id", coords = c("CenterX_global_px", "CenterY_global_px"))
patch_size4 <- merge(patch_size4, colData(ad_cms4)[match(patch_size4$patch_id, ad_cms4$patch_id),], by = "patch_id")

patch_size5 <- patchSize(ad_cms5, "patch_id", coords = c("CenterX_global_px", "CenterY_global_px"))
patch_size5 <- merge(patch_size5, colData(ad_cms5)[match(patch_size5$patch_id, ad_cms5$patch_id),], by = "patch_id")

plotSpatial(ad_cms1, node_color_by = "patch_id", img_id = "sample_ID", node_size_fix = 1, coords = c("CenterX_global_px", "CenterY_global_px")) + theme(legend.position = "none") + scale_color_manual(values = col_PD)
plotSpatial(ad_cms2, node_color_by = "patch_id", img_id = "sample_ID", node_size_fix = 1, coords = c("CenterX_global_px", "CenterY_global_px")) + theme(legend.position = "none") + scale_color_manual(values = col_PD)
plotSpatial(ad_cms3, node_color_by = "patch_id", img_id = "sample_ID", node_size_fix = 1, coords = c("CenterX_global_px", "CenterY_global_px")) + theme(legend.position = "none") + scale_color_manual(values = col_PD)
plotSpatial(ad_cms4, node_color_by = "patch_id", img_id = "sample_ID", node_size_fix = 1, coords = c("CenterX_global_px", "CenterY_global_px")) + theme(legend.position = "none") + scale_color_manual(values = col_PD)
plotSpatial(ad_cms5, node_color_by = "patch_id", img_id = "sample_ID", node_size_fix = 1, coords = c("CenterX_global_px", "CenterY_global_px")) + theme(legend.position = "none")

patch_size <- as.data.frame(patch_size)
patch_size2 <- as.data.frame(patch_size2)
patch_size3 <- as.data.frame(patch_size3)
patch_size4 <- as.data.frame(patch_size4)
patch_size5 <- as.data.frame(patch_size5)

patch_size_merge <- rbind(patch_size, patch_size2, patch_size3, patch_size4, patch_size5)

ggplot(patch_size_merge) + 
  geom_boxplot(aes(CMS, log10(size))) +
  geom_point(aes(CMS, log10(size))) + theme(axis.text.x = element_text(angle=90))

patch_size_merge <- patch_size_merge %>% select_if(~ !any(is.na(.)))
write_csv(patch_size_merge,"CMS_Tcell_patchsize.csv")

# MMR clusters
ad_dmmr <- patchDetection(ad_dmmr, patch_cells = ad_dmmr$metaclusters == "T cell", img_id = "sample_ID", expand_by = 60, min_patch_size = 3, #change this
                          coords = c("CenterX_global_px", "CenterY_global_px"), colPairName = "expansion_interaction_graph")
ad_pmmr <- patchDetection(ad_pmmr, patch_cells = ad_pmmr$metaclusters == "T cell", img_id = "sample_ID", expand_by = 60, min_patch_size = 3, #change this
                          coords = c("CenterX_global_px", "CenterY_global_px"), colPairName = "expansion_interaction_graph")
ad_partial <- patchDetection(ad_partial, patch_cells = ad_partial$metaclusters == "T cell", img_id = "sample_ID", expand_by = 60, min_patch_size = 3, #change this
                          coords = c("CenterX_global_px", "CenterY_global_px"), colPairName = "expansion_interaction_graph")

patch_size <- patchSize(ad_dmmr, "patch_id", coords = c("CenterX_global_px", "CenterY_global_px"))
patch_size <- merge(patch_size, colData(ad_dmmr)[match(patch_size$patch_id, ad_dmmr$patch_id),], by = "patch_id")

patch_size2 <- patchSize(ad_pmmr, "patch_id", coords = c("CenterX_global_px", "CenterY_global_px"))
patch_size2 <- merge(patch_size2, colData(ad_pmmr)[match(patch_size2$patch_id, ad_pmmr$patch_id),], by = "patch_id")

patch_size3 <- patchSize(ad_partial, "patch_id", coords = c("CenterX_global_px", "CenterY_global_px"))
patch_size3 <- merge(patch_size3, colData(ad_partial)[match(patch_size3$patch_id, ad_partial$patch_id),], by = "patch_id")

plotSpatial(ad_dmmr, node_color_by = "patch_id", img_id = "sample_ID", node_size_fix = 1, coords = c("CenterX_global_px", "CenterY_global_px")) + theme(legend.position = "none") + scale_color_manual(values = col_PD)
plotSpatial(ad_pmmr, node_color_by = "patch_id", img_id = "sample_ID", node_size_fix = 1, coords = c("CenterX_global_px", "CenterY_global_px")) + theme(legend.position = "none")
plotSpatial(ad_partial, node_color_by = "patch_id", img_id = "sample_ID", node_size_fix = 1, coords = c("CenterX_global_px", "CenterY_global_px")) + theme(legend.position = "none") + scale_color_manual(values = col_PD)

patch_size <- as.data.frame(patch_size)
patch_size2 <- as.data.frame(patch_size2)
patch_size3 <- as.data.frame(patch_size3)

patch_size_merge <- rbind(patch_size, patch_size2, patch_size3)

ggplot(patch_size_merge) + 
  geom_boxplot(aes(MMR_Status_Jen, log10(size))) +
  geom_point(aes(MMR_Status_Jen, log10(size))) + theme(axis.text.x = element_text(angle=90))

patch_size_merge <- patch_size_merge %>% select_if(~ !any(is.na(.)))
write_csv(patch_size_merge,"MMR_Tcell_patchsize.csv")

##################################################################################################################################################################################################################################################################################################################################################################################################################################################################

# 3- Directed interaction analysis 
# The imcRtools package further implements an interaction testing strategy proposed by (Schulz et al. 2018) 
 #where the hypothesis is tested if at least n cells of a certain type are located around a target cell type (from_cell). 
# This type of testing can be performed by selecting method = "patch" and specifying the number of patch cells via the patch_size parameter.

ep1 <- testInteractions(ad_ep1, group_by = "sample_ID", label = "CIPR_ordered_clusters", 
                         colPairName = "expansion_interaction_graph",method = "patch", patch_size = 3, BPPARAM = SerialParam(RNGseed = 221029))
ep2 <- testInteractions(ad_ep2, group_by = "sample_ID", label = "CIPR_ordered_clusters", 
                        colPairName = "expansion_interaction_graph",method = "patch", patch_size = 3, BPPARAM = SerialParam(RNGseed = 221029))
ep3 <- testInteractions(ad_ep3, group_by = "sample_ID", label = "CIPR_ordered_clusters", 
                        colPairName = "expansion_interaction_graph",method = "patch", patch_size = 3, BPPARAM = SerialParam(RNGseed = 221029))

tme1 <- testInteractions(ad_tme1, group_by = "sample_ID", label = "CIPR_ordered_clusters", 
                        colPairName = "expansion_interaction_graph",method = "patch", patch_size = 3, BPPARAM = SerialParam(RNGseed = 221029))
tme2 <- testInteractions(ad_tme2, group_by = "sample_ID", label = "CIPR_ordered_clusters", 
                        colPairName = "expansion_interaction_graph",method = "patch", patch_size = 3, BPPARAM = SerialParam(RNGseed = 221029))
tme3 <- testInteractions(ad_tme3, group_by = "sample_ID", label = "CIPR_ordered_clusters", 
                        colPairName = "expansion_interaction_graph",method = "patch", patch_size = 3, BPPARAM = SerialParam(RNGseed = 221029))

sma1 <- testInteractions(ad_sma1, group_by = "sample_ID", label = "CIPR_ordered_clusters", 
                        colPairName = "expansion_interaction_graph",method = "patch", patch_size = 3, BPPARAM = SerialParam(RNGseed = 221029))
sma2 <- testInteractions(ad_sma2, group_by = "sample_ID", label = "CIPR_ordered_clusters", 
                        colPairName = "expansion_interaction_graph",method = "patch", patch_size = 3, BPPARAM = SerialParam(RNGseed = 221029))
sma3 <- testInteractions(ad_sma3, group_by = "sample_ID", label = "CIPR_ordered_clusters", 
                        colPairName = "expansion_interaction_graph",method = "patch", patch_size = 3, BPPARAM = SerialParam(RNGseed = 221029))

cms1 <- testInteractions(ad_cms1, group_by = "sample_ID", label = "CIPR_ordered_clusters", 
                        colPairName = "expansion_interaction_graph",method = "patch", patch_size = 3, BPPARAM = SerialParam(RNGseed = 221029))
cms2 <- testInteractions(ad_cms2, group_by = "sample_ID", label = "CIPR_ordered_clusters", 
                        colPairName = "expansion_interaction_graph",method = "patch", patch_size = 3, BPPARAM = SerialParam(RNGseed = 221029))
cms3 <- testInteractions(ad_cms3, group_by = "sample_ID", label = "CIPR_ordered_clusters", 
                        colPairName = "expansion_interaction_graph",method = "patch", patch_size = 3, BPPARAM = SerialParam(RNGseed = 221029))
cms4 <- testInteractions(ad_cms4, group_by = "sample_ID", label = "CIPR_ordered_clusters", 
                         colPairName = "expansion_interaction_graph",method = "patch", patch_size = 3, BPPARAM = SerialParam(RNGseed = 221029))
cms5 <- testInteractions(ad_cms5, group_by = "sample_ID", label = "CIPR_ordered_clusters", 
                         colPairName = "expansion_interaction_graph",method = "patch", patch_size = 3, BPPARAM = SerialParam(RNGseed = 221029))

dmmr <- testInteractions(ad_dmmr, group_by = "sample_ID", label = "CIPR_ordered_clusters", 
                         colPairName = "expansion_interaction_graph",method = "patch", patch_size = 3, BPPARAM = SerialParam(RNGseed = 221029))
pmmr <- testInteractions(ad_pmmr, group_by = "sample_ID", label = "CIPR_ordered_clusters", 
                         colPairName = "expansion_interaction_graph",method = "patch", patch_size = 3, BPPARAM = SerialParam(RNGseed = 221029))
partial <- testInteractions(ad_partial, group_by = "sample_ID", label = "CIPR_ordered_clusters", 
                         colPairName = "expansion_interaction_graph",method = "patch", patch_size = 3, BPPARAM = SerialParam(RNGseed = 221029))

# recurrences - directed NE
norec <- testInteractions(ad_norecurrence, group_by = "sample_ID", label = "CIPR_ordered_clusters", 
                         colPairName = "expansion_interaction_graph",method = "patch", patch_size = 3, BPPARAM = SerialParam(RNGseed = 221029))
local <- testInteractions(ad_local, group_by = "sample_ID", label = "CIPR_ordered_clusters", 
                         colPairName = "expansion_interaction_graph",method = "patch", patch_size = 3, BPPARAM = SerialParam(RNGseed = 221029))
liver <- testInteractions(ad_liver, group_by = "sample_ID", label = "CIPR_ordered_clusters", 
                         colPairName = "expansion_interaction_graph",method = "patch", patch_size = 3, BPPARAM = SerialParam(RNGseed = 221029))
brain <- testInteractions(ad_brain, group_by = "sample_ID", label = "CIPR_ordered_clusters", 
                         colPairName = "expansion_interaction_graph",method = "patch", patch_size = 3, BPPARAM = SerialParam(RNGseed = 221029))
multisite <- testInteractions(ad_multisite, group_by = "sample_ID", label = "CIPR_ordered_clusters", 
                         colPairName = "expansion_interaction_graph",method = "patch", patch_size = 3, BPPARAM = SerialParam(RNGseed = 221029))

# recurrences - undirected NE
norec2 <- testInteractions(ad_norecurrence, group_by = "sample_ID", label = "CIPR_ordered_clusters", 
                          colPairName = "expansion_interaction_graph", BPPARAM = SerialParam(RNGseed = 221029))
local2 <- testInteractions(ad_local, group_by = "sample_ID", label = "CIPR_ordered_clusters", 
                          colPairName = "expansion_interaction_graph", BPPARAM = SerialParam(RNGseed = 221029))
liver2 <- testInteractions(ad_liver, group_by = "sample_ID", label = "CIPR_ordered_clusters", 
                          colPairName = "expansion_interaction_graph", BPPARAM = SerialParam(RNGseed = 221029))
brain2 <- testInteractions(ad_brain, group_by = "sample_ID", label = "CIPR_ordered_clusters", 
                          colPairName = "expansion_interaction_graph", BPPARAM = SerialParam(RNGseed = 221029))
multisite2 <- testInteractions(ad_multisite, group_by = "sample_ID", label = "CIPR_ordered_clusters", 
                              colPairName = "expansion_interaction_graph",method = "patch", patch_size = 3, BPPARAM = SerialParam(RNGseed = 221029))

library(scales)
ep1 %>% as_tibble() %>%
  group_by(from_label, to_label) %>%
  summarize(mean_sigval = mean(sigval, na.rm = TRUE)) %>%
  ggplot() +
  geom_tile(aes(from_label, to_label, fill = mean_sigval)) +
  scale_fill_gradient2(low = muted("blue"), mid = "white", high = muted("red")) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

ep2 %>% as_tibble() %>%
  group_by(from_label, to_label) %>%
  summarize(mean_sigval = mean(sigval, na.rm = TRUE)) %>%
  ggplot() +
  geom_tile(aes(from_label, to_label, fill = mean_sigval)) +
  scale_fill_gradient2(low = muted("blue"), mid = "white", high = muted("red")) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

ep3 %>% as_tibble() %>%
  group_by(from_label, to_label) %>%
  summarize(mean_sigval = mean(sigval, na.rm = TRUE)) %>%
  ggplot() +
  geom_tile(aes(from_label, to_label, fill = mean_sigval)) +
  scale_fill_gradient2(low = muted("blue"), mid = "white", high = muted("red")) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

tme1 %>% as_tibble() %>%
  group_by(from_label, to_label) %>%
  summarize(mean_sigval = mean(sigval, na.rm = TRUE)) %>%
  ggplot() +
  geom_tile(aes(from_label, to_label, fill = mean_sigval)) +
  scale_fill_gradient2(low = muted("blue"), mid = "white", high = muted("red")) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

tme2 %>% as_tibble() %>%
  group_by(from_label, to_label) %>%
  summarize(mean_sigval = mean(sigval, na.rm = TRUE)) %>%
  ggplot() +
  geom_tile(aes(from_label, to_label, fill = mean_sigval)) +
  scale_fill_gradient2(low = muted("blue"), mid = "white", high = muted("red")) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

tme3 %>% as_tibble() %>%
  group_by(from_label, to_label) %>%
  summarize(mean_sigval = mean(sigval, na.rm = TRUE)) %>%
  ggplot() +
  geom_tile(aes(from_label, to_label, fill = mean_sigval)) +
  scale_fill_gradient2(low = muted("blue"), mid = "white", high = muted("red")) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

sma1 %>% as_tibble() %>%
  group_by(from_label, to_label) %>%
  summarize(mean_sigval = mean(sigval, na.rm = TRUE)) %>%
  ggplot() +
  geom_tile(aes(from_label, to_label, fill = mean_sigval)) +
  scale_fill_gradient2(low = muted("blue"), mid = "white", high = muted("red")) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

sma2 %>% as_tibble() %>%
  group_by(from_label, to_label) %>%
  summarize(mean_sigval = mean(sigval, na.rm = TRUE)) %>%
  ggplot() +
  geom_tile(aes(from_label, to_label, fill = mean_sigval)) +
  scale_fill_gradient2(low = muted("blue"), mid = "white", high = muted("red")) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

sma3 %>% as_tibble() %>%
  group_by(from_label, to_label) %>%
  summarize(mean_sigval = mean(sigval, na.rm = TRUE)) %>%
  ggplot() +
  geom_tile(aes(from_label, to_label, fill = mean_sigval)) +
  scale_fill_gradient2(low = muted("blue"), mid = "white", high = muted("red")) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

cms1 %>% as_tibble() %>%
  group_by(from_label, to_label) %>%
  summarize(mean_sigval = mean(sigval, na.rm = TRUE)) %>%
  ggplot() +
  geom_tile(aes(from_label, to_label, fill = mean_sigval)) +
  scale_fill_gradient2(low = muted("blue"), mid = "white", high = muted("red")) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

cms2 %>% as_tibble() %>%
  group_by(from_label, to_label) %>%
  summarize(mean_sigval = mean(sigval, na.rm = TRUE)) %>%
  ggplot() +
  geom_tile(aes(from_label, to_label, fill = mean_sigval)) +
  scale_fill_gradient2(low = muted("blue"), mid = "white", high = muted("red")) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

cms3 %>% as_tibble() %>%
  group_by(from_label, to_label) %>%
  summarize(mean_sigval = mean(sigval, na.rm = TRUE)) %>%
  ggplot() +
  geom_tile(aes(from_label, to_label, fill = mean_sigval)) +
  scale_fill_gradient2(low = muted("blue"), mid = "white", high = muted("red")) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

cms4 %>% as_tibble() %>%
  group_by(from_label, to_label) %>%
  summarize(mean_sigval = mean(sigval, na.rm = TRUE)) %>%
  ggplot() +
  geom_tile(aes(from_label, to_label, fill = mean_sigval)) +
  scale_fill_gradient2(low = muted("blue"), mid = "white", high = muted("red")) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

cms5 %>% as_tibble() %>%
  group_by(from_label, to_label) %>%
  summarize(mean_sigval = mean(sigval, na.rm = TRUE)) %>%
  ggplot() +
  geom_tile(aes(from_label, to_label, fill = mean_sigval)) +
  scale_fill_gradient2(low = muted("blue"), mid = "white", high = muted("red")) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

dmmr %>% as_tibble() %>%
  group_by(from_label, to_label) %>%
  summarize(mean_sigval = mean(sigval, na.rm = TRUE)) %>%
  ggplot() +
  geom_tile(aes(from_label, to_label, fill = mean_sigval)) +
  scale_fill_gradient2(low = muted("blue"), mid = "white", high = muted("red")) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

pmmr %>% as_tibble() %>%
  group_by(from_label, to_label) %>%
  summarize(mean_sigval = mean(sigval, na.rm = TRUE)) %>%
  ggplot() +
  geom_tile(aes(from_label, to_label, fill = mean_sigval)) +
  scale_fill_gradient2(low = muted("blue"), mid = "white", high = muted("red")) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

partial %>% as_tibble() %>%
  group_by(from_label, to_label) %>%
  summarize(mean_sigval = mean(sigval, na.rm = TRUE)) %>%
  ggplot() +
  geom_tile(aes(from_label, to_label, fill = mean_sigval)) +
  scale_fill_gradient2(low = muted("blue"), mid = "white", high = muted("red")) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

norec2 %>% as_tibble() %>%
  group_by(from_label, to_label) %>%
  summarize(sum_sigval = sum(sigval, na.rm = TRUE)) %>%
  ggplot() +
  geom_tile(aes(from_label, to_label, fill = sum_sigval)) +
  scale_fill_gradient2(low = muted("blue"), mid = "white", high = muted("red")) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

local2 %>% as_tibble() %>%
  group_by(from_label, to_label) %>%
  summarize(sum_sigval = sum(sigval, na.rm = TRUE)) %>%
  ggplot() +
  geom_tile(aes(from_label, to_label, fill = sum_sigval)) +
  scale_fill_gradient2(low = muted("blue"), mid = "white", high = muted("red")) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

liver2 %>% as_tibble() %>%
  group_by(from_label, to_label) %>%
  summarize(sum_sigval = sum(sigval, na.rm = TRUE)) %>%
  ggplot() +
  geom_tile(aes(from_label, to_label, fill = sum_sigval)) +
  scale_fill_gradient2(low = muted("blue"), mid = "white", high = muted("red")) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

brain %>% as_tibble() %>%
  group_by(from_label, to_label) %>%
  summarize(sum_sigval = sum(sigval, na.rm = TRUE)) %>%
  ggplot() +
  geom_tile(aes(from_label, to_label, fill = sum_sigval)) +
  scale_fill_gradient2(low = muted("blue"), mid = "white", high = muted("red")) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

multisite %>% as_tibble() %>%
  group_by(from_label, to_label) %>%
  summarize(sum_sigval = sum(sigval, na.rm = TRUE)) %>%
  ggplot() +
  geom_tile(aes(from_label, to_label, fill = sum_sigval)) +
  scale_fill_gradient2(low = muted("blue"), mid = "white", high = muted("red")) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))


#These results are comparable to the interaction testing presented above. 
#The main difference comes from the lack of symmetry. We can now for example see that 3 or more myeloid cells sit around CD4 T cells 
#while this interaction is not as strong when considering CD4 T cells sitting around myeloid cells.

#Finally, we save the updated SpatialExperiment object.
saveRDS(spe, "spe.rds")
